Patch 6.1a.031 Problem:    Accessing the 'balloondelay' variable may cause a crash. Solution:   Make the variable for 'balloondelay' a long. (Olaf Seibert) Files:	    src/option.h

+++ src/option.h
@@ -292,7 +292,7 @@ EXTERN char_u	*p_bex;		/* 'backupext' */
 EXTERN char_u	*p_bsk;		/* 'backupskip' */
 #endif
 #ifdef FEAT_BEVAL
-EXTERN int	p_bdlay;	/* 'balloondelay' */
+EXTERN long	p_bdlay;	/* 'balloondelay' */
 #endif
 #ifdef FEAT_SUN_WORKSHOP
 EXTERN int	p_beval;	/* 'ballooneval' */
